Разбиение файлов на классы
