+++ rovmodel/rovmodel.pro
@@ -4,8 +4,14 @@ CONFIG -= app_bundle
 CONFIG -= qt
 
 SOURCES += \
+    floatingobject.cpp \
+    core.cpp \
     robot.cpp
 
 include(deployment.pri)
 qtcAddDeployment()
+
+HEADERS += \
+    floatingobject.h \
+    robot.h
